Cluster and Cloud Computing Project/Exploring Patterns Data/Mastadon/requirements.txt
Mastodon.py
couchdb
